+++ Whatnow-Domain/src/main/kotlin/com/depromeet/whatnow/events/domainEvent/InteractionHistoryRegisterEvent.kt
@@ -1,4 +1,4 @@
-package com.depromeet.whatnow.domains.interactionhistory.event
+package com.depromeet.whatnow.events.domainEvent
 
 import com.depromeet.whatnow.common.aop.event.DomainEvent
 import com.depromeet.whatnow.domains.interaction.domain.InteractionType
